+++ tests/unit/resources/test_model.py
@@ -74,7 +74,8 @@ class TestModels(BaseTestCase):
             'actions': {
                 'GetFrobs': {
                     'resource': {
-                        'type': 'Frob'
+                        'type': 'Frob',
+                        'path': 'Container.Frobs[]'
                     }
                 }
             }
@@ -84,6 +85,7 @@ class TestModels(BaseTestCase):
 
         action = model.actions[0]
         self.assertEqual(action.resource.type, 'Frob')
+        self.assertEqual(action.resource.path, 'Container.Frobs[]')
         self.assertIsInstance(action.resource.model, ResourceModel)
         self.assertEqual(action.resource.model.name, 'Frob')
 
@@ -144,7 +146,7 @@ class TestModels(BaseTestCase):
 
     def test_resource_references(self):
         model_def = {
-            'hasOne': {
+            'belongsTo': {
                 'Frob': {
                     'resource': {
                         'type': 'Frob',
@@ -152,13 +154,6 @@ class TestModels(BaseTestCase):
                             {'target':'Id', 'sourceType':'dataMember',
                              'source':'FrobId'}
                         ]
-                    },
-                }
-            },
-            'hasSome': {
-                'Frobs': {
-                    'resource': {
-                        'type': 'Frob'
                     }
                 }
             }
@@ -169,7 +164,7 @@ class TestModels(BaseTestCase):
         model = ResourceModel('test', model_def, resource_defs)
 
         self.assertIsInstance(model.references, list)
-        self.assertEqual(len(model.references), 2)
+        self.assertEqual(len(model.references), 1)
 
         ref = model.references[0]
         self.assertEqual(ref.name, 'Frob')
@@ -178,11 +173,6 @@ class TestModels(BaseTestCase):
         self.assertEqual(ref.resource.identifiers[0].source_type,
                          'dataMember')
         self.assertEqual(ref.resource.identifiers[0].source, 'FrobId')
-
-        ref2 = model.references[1]
-        self.assertEqual(ref2.name, 'Frobs')
-        self.assertEqual(ref2.resource.type, 'Frob')
-        self.assertEqual(len(ref2.resource.identifiers), 0)
 
     def test_reverse_reference(self):
         # Here the Code resource has no explicit ``hasOne`` defined, however